Move the AutofacStartable into AutofacServiceRegistration
