+++ Source/EventFlow/Configuration/Registrations/AutofacServiceRegistration.cs
@@ -21,8 +21,13 @@
 // CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 
 using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Threading;
+using System.Threading.Tasks;
 using Autofac;
 using EventFlow.Configuration.Decorators;
+using EventFlow.Core;
 using EventFlow.Extensions;
 
 namespace EventFlow.Configuration.Registrations
@@ -174,6 +179,50 @@ namespace EventFlow.Configuration.Registrations
             }
 
             return new AutofacRootResolver(container);
+        }
+
+        public class AutofacStartable : IStartable
+        {
+            private readonly IReadOnlyCollection<IBootstrap> _bootstraps;
+
+            public AutofacStartable(
+                IEnumerable<IBootstrap> bootstraps)
+            {
+                _bootstraps = OrderBootstraps(bootstraps);
+            }
+
+            public void Start()
+            {
+                using (var a = AsyncHelper.Wait)
+                {
+                    a.Run(StartAsync(CancellationToken.None));
+                }
+            }
+
+            private Task StartAsync(CancellationToken cancellationToken)
+            {
+                return Task.WhenAll(_bootstraps.Select(b => b.BootAsync(cancellationToken)));
+            }
+
+            private static IReadOnlyCollection<IBootstrap> OrderBootstraps(IEnumerable<IBootstrap> bootstraps)
+            {
+                var list = bootstraps
+                    .Select(b => new
+                    {
+                        Bootstrap = b,
+                        AssemblyName = b.GetType().Assembly.GetName().Name,
+                    })
+                    .ToList();
+                var eventFlowBootstraps = list
+                    .Where(a => a.AssemblyName.StartsWith("EventFlow"))
+                    .OrderBy(a => a.AssemblyName)
+                    .Select(a => a.Bootstrap);
+                var otherBootstraps = list
+                    .Where(a => !a.AssemblyName.StartsWith("EventFlow"))
+                    .OrderBy(a => a.AssemblyName)
+                    .Select(a => a.Bootstrap);
+                return eventFlowBootstraps.Concat(otherBootstraps).ToList();
+            }
         }
     }
 }
